var/requirements.txt
# VAR Analysis System - Requirements
# Install with: pip install -r requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.8.0
chromadb>=0.5.0
python-multipart==0.0.6
pydantic>=2.9.0
pypdf>=3.17.0
python-dotenv==1.0.0
slowapi>=0.1.9
